chart関連の型移動
