--- common/types/chart.ts
@@ -1,8 +1,0 @@
-export type ChartDataProps = {
-  labels: string[];
-  label: string;
-  data: number[];
-  bgColor: string[];
-  bdColor: string[];
-  bdWidth: number;
-}
