+++ common/types/utils.ts
@@ -1,6 +1,5 @@
 import { ReactNode } from 'react';
 import { StaticImageData } from 'next/image';
-import { TooltipItem } from 'chart.js';
 
 export type ChildrenProps = {
   children: ReactNode;
@@ -20,71 +19,6 @@ export type sidebarMenus = {
   label: string;
   value: string;
   icon: React.JSX.Element;
-};
-
-export type ChartData = {
-  labels: string[];
-  datasets: {
-    label: string;
-    data: number[];
-    bgColor: string[];
-    bdColor: string[];
-    bdWidth: number;
-  };
-};
-
-export type AnnotationLine = {
-  type: 'line';
-  yMin: number;
-  yMax: number;
-  borderColor: string;
-  borderWidth: number;
-};
-
-export type ChartOption = {
-  responsive: boolean;
-  plugins: {
-    title: {
-      display: boolean;
-      text: string;
-    };
-    tooltip?: {
-      callbacks: {
-        label: (context: TooltipItem<'bar'>) => string;
-      };
-    };
-    annotation?: {
-      annotations: {
-        legendLine: AnnotationLine;
-        masterLine: AnnotationLine;
-        diamondLine: AnnotationLine;
-        platinumLine: AnnotationLine;
-        goldLine: AnnotationLine;
-        silverLine: AnnotationLine;
-      };
-    };
-  };
-  scales?: {
-    y: {
-      suggestedMax: number;
-      title: {
-        display: boolean;
-        text: string;
-      };
-    };
-    x: {
-      title: {
-        display: boolean;
-        text: string;
-      }
-    }
-  };
-};
-
-export type ChartProps = {
-  data: any;
-  option: any;
-  type: 'bar' | 'pie';
 };
 
 export type FormSubmitButtonProps = {
